+++ src/Network/Ethereum/Web3.hs
@@ -24,9 +24,7 @@ module Network.Ethereum.Web3 (
   , runWeb3
 
   -- ** Basic transaction sending
-  , Account(..)
-  , withTarget
-  , value
+  , module Network.Ethereum.Account
 
   -- ** Basic event listening
   , EventAction(..)
@@ -42,10 +40,9 @@ module Network.Ethereum.Web3 (
   ) where
 
 import           Data.Solidity.Prim
-import           Network.Ethereum.Account        (Account (..))
+import           Network.Ethereum.Account
 import           Network.Ethereum.Api.Provider   (Web3, runWeb3)
 import           Network.Ethereum.Api.Types      (Call (..))
-import           Network.Ethereum.Contract       (value, withTarget)
 import           Network.Ethereum.Contract.Event (EventAction (..), event,
                                                   event')
 import           Network.Ethereum.Unit